Move out app-related slides for later use
